refactor: unify minibus.DropExcess into concurrent.BreakBackpressure
minibus.DropExcess and concurrent.BreakBackpressure were byte-for-byte identical. Remove the minibus copy and point its callers at the concurrent package, which already owns the buffered variant and a more descriptive name. The DropExcess test moves across too, giving the concurrent backpressure helpers test coverage they previously lacked.
