--- pkg/minibus/util.go
@@ -1,37 +1,0 @@
-package minibus
-
-// DropExcess will wrap a channel to break backpressure between its input and output.
-// It will drop incoming messages when the consumer can't process them fast enough.
-// When the consumer receives, it will always get the most recent message sent by the producer.
-func DropExcess[T any](in <-chan T) <-chan T {
-	out := make(chan T)
-	go func() {
-		defer close(out)
-		var message T
-		hasMessage := false
-		for {
-			if hasMessage {
-				select {
-				case newMessage, ok := <-in:
-					if !ok {
-						return
-					}
-					// replace the buffered message, discarding the old one
-					message = newMessage
-
-				case out <- message:
-					// message sent successfully
-					hasMessage = false
-				}
-			} else {
-				newMessage, ok := <-in
-				if !ok {
-					return
-				}
-				message = newMessage
-				hasMessage = true
-			}
-		}
-	}()
-	return out
-}
